Adding location routes and structure for backend
